[IMP] account: add compute_fiscalyear_dates method on res.company
[IMP] account: add compute_fiscalyear_dates method on res.company

The following commits
  https://github.com/odoo/odoo/commit/f5e746d0d9768958ed18faebd4dceb3c37fc08bf
  https://github.com/odoo-dev/enterprise/commit/ed383522791276b17ba68a4f0d08f9e3ed4e783c

move the  compute_fiscalyear_dates method from the community account module to the enterprise one account_accounting.

This commit ensures that the method can be used when account_accounting is not installed.
As the fiscal year is irrelevant when account_accounting is not installed, this commit displays the calendar year.

taskid: 2329108
See also: taskid 220194

closes odoo/odoo#60140

X-original-commit: 544cb3682033e9f644b48b97dd5197eaf9282ff5
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>